Separated Usage tab into component
